+++ cmd/deploy.go
@@ -66,14 +66,15 @@ func runDeployCmd(cmd *cobra.Command, args []string) {
 			}
 		})
 	}
-	validators.PerformGkeVulnerabilitiesCheck(cmd, args)
-	doDeploy(deplRoot)
+	skipSecurity, _ := cmd.Flags().GetBool("skip-gke-security-check")
+	doDeploy(deplRoot, skipSecurity)
 }
 
-func doDeploy(deplRoot string) {
+func doDeploy(deplRoot string, skipSecurity bool) {
 	artDir := getArtifactsDir(deplRoot)
 	checkErr(shell.CheckWritableDir(artDir), nil)
 	bp, ctx := artifactBlueprintOrDie(artDir)
+	validators.PerformGkeVulnerabilitiesCheck(skipSecurity, &bp)
 	groups := bp.Groups
 	checkErr(validateGroupSelectionFlags(bp), ctx)
 	checkErr(validateRuntimeDependencies(deplRoot, groups), ctx)
