+++ pkg/config/config_test.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -33,10 +32,8 @@ import (
 
 	"hpc-toolkit/pkg/modulereader"
 
-	"github.com/google/go-cmp/cmp"
 	"github.com/pkg/errors"
 	"github.com/zclconf/go-cty/cty"
-	"google.golang.org/api/option"
 	. "gopkg.in/check.v1"
 )
 
@@ -1466,220 +1463,6 @@ func TestGetKeyFromBlueprint(t *testing.T) {
 
 			if actual != tt.expected {
 				t.Errorf("getKeyFromBlueprint(%q) = %q, want %q", tt.key, actual, tt.expected)
-			}
-		})
-	}
-}
-
-func mockFetchGKEVersionForPrefix(projectID, region, prefix, releaseChannel string) (string, error) {
-	// Mock responses based on channel and prefix
-	if projectID == "test-project" && region == "us-central1" {
-		if releaseChannel == "RAPID" && strings.HasPrefix("1.36.", prefix) {
-			return "1.36.1-gke.100", nil
-		}
-		if releaseChannel == "REGULAR" && strings.HasPrefix("1.35.", prefix) {
-			return "1.35.2-gke.200", nil
-		}
-		if prefix == "1.34." {
-			return "1.34.9-gke.100", nil // Fallback test
-		}
-	}
-	return "", nil
-}
-
-// In TestResolveGKEVersions
-func TestResolveGKEVersions(t *testing.T) {
-	// Temporarily override the function pointer for testing
-	originalFetchFunc := fetchGKEVersionFunc
-	fetchGKEVersionFunc = mockFetchGKEVersionForPrefix
-	// Restore original function after test
-	defer func() { fetchGKEVersionFunc = originalFetchFunc }()
-
-	testCases := []struct {
-		name      string
-		blueprint Blueprint
-		expected  []string
-		hasError  bool
-	}{
-		{
-			name: "simple min_master_version",
-			blueprint: Blueprint{
-				Vars: NewDict(map[string]cty.Value{
-					"project_id": cty.StringVal("test-project"),
-					"region":     cty.StringVal("us-central1"),
-				}),
-				Groups: []Group{
-					{
-						Name: "primary",
-						Modules: []Module{
-							{
-								ID:     "gke",
-								Source: "modules/scheduler/gke-cluster",
-								Settings: NewDict(map[string]cty.Value{
-									"min_master_version": cty.StringVal("1.37.0"),
-								}),
-							},
-						},
-					},
-				},
-			},
-			expected: []string{"1.37.0"},
-		},
-		{
-			name: "version_prefix rapid",
-			blueprint: Blueprint{
-				Vars: NewDict(map[string]cty.Value{
-					"project_id": cty.StringVal("test-project"),
-					"region":     cty.StringVal("us-central1"),
-				}),
-				Groups: []Group{
-					{
-						Name: "primary",
-						Modules: []Module{
-							{
-								ID:     "gke",
-								Source: "modules/scheduler/gke-cluster",
-								Settings: NewDict(map[string]cty.Value{
-									"version_prefix":  cty.StringVal("1.36."),
-									"release_channel": cty.StringVal("RAPID"),
-								}),
-							},
-						},
-					},
-				},
-			},
-			expected: []string{"1.36.1-gke.100"},
-		},
-		{
-			name: "version_prefix regular default",
-			blueprint: Blueprint{
-				Vars: NewDict(map[string]cty.Value{
-					"project_id": cty.StringVal("test-project"),
-					"region":     cty.StringVal("us-central1"),
-				}),
-				Groups: []Group{
-					{
-						Name: "primary",
-						Modules: []Module{
-							{
-								ID:     "gke",
-								Source: "modules/scheduler/gke-cluster",
-								Settings: NewDict(map[string]cty.Value{
-									"version_prefix":  cty.StringVal("1.35."),
-									"release_channel": cty.StringVal("REGULAR"),
-								}),
-							},
-						},
-					},
-				},
-			},
-			expected: []string{"1.35.2-gke.200"},
-		},
-		{
-			name: "version_prefix fallback",
-			blueprint: Blueprint{
-				Vars: NewDict(map[string]cty.Value{
-					"project_id": cty.StringVal("test-project"),
-					"region":     cty.StringVal("us-central1"),
-				}),
-				Groups: []Group{
-					{
-						Name: "primary",
-						Modules: []Module{
-							{
-								ID:     "gke",
-								Source: "modules/scheduler/gke-cluster",
-								Settings: NewDict(map[string]cty.Value{
-									"version_prefix": cty.StringVal("1.32."), // Non-existent in mock
-								}),
-							},
-						},
-					},
-				},
-			},
-			expected: []string{},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			versions, err := ResolveGKEVersions(&tc.blueprint)
-			if (err != nil) != tc.hasError {
-				t.Errorf("ResolveGKEVersions() error = %v, wantErr %v", err, tc.hasError)
-				return
-			}
-
-			if diff := cmp.Diff(tc.expected, versions); diff != "" {
-				t.Errorf("ResolveGKEVersions() mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestFetchLatestGKEVersionForPrefix(t *testing.T) {
-	// 1. Create a mock HTTP server simulating the GCP Container API Response
-	mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Verify the API URL path is correctly formatted
-		expectedPath := "/v1/projects/test-project/locations/us-central1/serverConfig"
-		if r.URL.Path != expectedPath {
-			t.Errorf("Expected path %q, got %q", expectedPath, r.URL.Path)
-		}
-
-		// Provide a mock JSON response matching container.ServerConfig
-		w.Header().Set("Content-Type", "application/json")
-		fmt.Fprintln(w, `{
-			"channels": [
-				{
-					"channel": "RAPID",
-					"validVersions": ["1.35.3-gke.1943000", "1.36.0-gke.1555000"]
-				},
-				{
-					"channel": "REGULAR",
-					"validVersions": ["1.34.7-gke.1292000", "1.35.1-gke.1000"]
-				}
-			]
-		}`)
-	}))
-	defer mockServer.Close()
-
-	// 2. Pass options to route traffic to the mock server and disable auth
-	clientOpt := option.WithEndpoint(mockServer.URL)
-	noAuthOpt := option.WithoutAuthentication()
-
-	tests := []struct {
-		name        string
-		prefix      string
-		channel     string
-		wantVersion string
-	}{
-		{
-			name:        "Highest version across all channels for 1.35 prefix",
-			prefix:      "1.35.",
-			channel:     "RAPID",
-			wantVersion: "1.35.3-gke.1943000", // Matches RAPID, which is higher than REGULAR's 1.35.1
-		},
-		{
-			name:        "Highest version for 1.34 prefix",
-			prefix:      "1.34.",
-			channel:     "REGULAR",
-			wantVersion: "1.34.7-gke.1292000",
-		},
-		{
-			name:        "Fallback to empty if no matching prefix is found",
-			prefix:      "1.29.",
-			channel:     "UNSPECIFIED",
-			wantVersion: "",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			got, err := fetchLatestGKEVersionForPrefix("test-project", "us-central1", tc.prefix, tc.channel, clientOpt, noAuthOpt)
-			if err != nil {
-				t.Fatalf("Unexpected error: %v", err)
-			}
-			if got != tc.wantVersion {
-				t.Errorf("fetchLatestGKEVersionForPrefix() = %v, want %v", got, tc.wantVersion)
 			}
 		})
 	}
